Major Update
many extensions to rtcwhq.py (see Readme)

+++ makepk3.py
@@ -3,9 +3,10 @@ import sys
 import zipfile
 import shutil
 
-# delete folder after zipping (True or False)
-delete=False
+# variables
+delete=False      # delete folder after zipping (True or False)
 
+# add files/folders to a zip archive
 def zipfolder(path, relname, archive):
 	paths = os.listdir(path)
 	for p in paths:
@@ -19,6 +20,7 @@ def zipfolder(path, relname, archive):
 		else:
 			archive.write(p1, p2) 
 
+# create a zip archive
 def create_zip(path, relname, archname):
 	archive = zipfile.ZipFile(archname, "w", zipfile.ZIP_DEFLATED)
 	if os.path.isdir(path):
@@ -29,11 +31,19 @@ def create_zip(path, relname, archname):
 	
 path = sys.argv[1]
 paths = os.listdir(path)
+
+print("----------------------------------------------------------------------")
+print("Creating PK3s from the folders in the root of "+path)
+print("----------------------------------------------------------------------")
+
 for p in paths:
+	
+	# take only the root folder inside the path
 	pk = os.path.join("", p)
-	print("----------------------------------------------------------------------")
-	print("Moving files to " + pk + ".pk3")
-	print("----------------------------------------------------------------------")
-	create_zip(path + "/" + pk,"",path + "/" + pk + "_HQ.pk3")
-	if(delete):
-		shutil.rmtree(path + "/" + pk, ignore_errors = True)
+
+	# only ZIP folders
+	if(os.path.isdir(path+"/"+pk)):
+		print("Moving files to " + pk + ".pk3")
+		create_zip(path + "/" + pk,"",path + "/" + pk + "_HQ.pk3")
+		if(delete):
+			shutil.rmtree(path + "/" + pk, ignore_errors = True)
